requirements.txt
streamlit
torch
torchvision
opencv-python-headless
scikit-learn
matplotlib
numpy
Pillow